backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.1.2
SQLAlchemy==2.0.29
alembic==1.13.1
